+++ RELEASE_NOTES.md
@@ -8,7 +8,10 @@
      * Chart.SingleStack: Basically Chart.Grid with one Column
  * Add Chart.withColorBar and Chart.withColorBarStyle to change the appearance of colorbars
  * [Add Sunburst Chart]()
+ * [Add Treemap Chart]()
  * [Add OHLC Chart]()
+ * [Add option to render all charts derived from the scatter trace type via WebGL as scattergl trace]()
+ * [Add Stackgroup related parameters to all charts derived from the scatter trace]()
 
 ### 1.2.2 - Apr 9 2020
  * [Opening Charts is now more or less OS agnostic](https://github.com/muehlhaus/FSharp.Plotly/commit/f6e3dceade085e43e7e56b478b9cf7b533a4fe55)